+++ database/migrations/6_create_transactions_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->date("date_operation")->nullable();
             $table->time("heure_operation")->nullable();
             $table->dateTime("date_end_trans")->nullable();
-            $table->integer("service_id");
+            $table->bigInteger("service_id");
            // $table->foreignId("service_id")->references("id")->on("services");
             $table->double("balance_before")->default(0);
             $table->double("debit")->default(0);
